refactor: move actual app theme logic calculating to useIdDark comp
