+++ src/composables/useAppThemeController.ts
@@ -1,15 +1,8 @@
-import { computed, watchEffect } from "vue"
-import { storeToRefs } from "pinia"
-import { usePreferredDark } from "@vueuse/core"
-import { useSettingsStore } from "@/stores/settings"
+import { watchEffect } from "vue"
+import { useIsDark } from "./useIsDark"
 
 export function useAppThemeController() {
-  const settings = useSettingsStore()
-  const { theme } = storeToRefs(settings)
-  const isDarkPreferred = usePreferredDark()
-  const isDark = computed(() =>
-    theme.value === "system" ? isDarkPreferred.value : theme.value === "dark"
-  )
+  const isDark = useIsDark()
   const root = document.querySelector(":root")
 
   watchEffect(() => {
